requirements.txt
contourpy==1.1.0
cycler==0.11.0
fonttools==4.42.1
kiwisolver==1.4.5
matplotlib==3.7.3
numpy==1.25.2
packaging==23.1
Pillow==10.0.0
pyparsing==3.1.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
pyqtgraph==0.13.3
pyserial==3.5
python-dateutil==2.8.2
six==1.16.0
